Fixing comparison in unit test

+++ tests/unit/general/test_arrays_by_name.cpp
@@ -124,7 +124,7 @@ TEST_CASE("ArraysByName Various Methods", "[ArraysByName]")
 
    // Get set names and verify that they are valid
    std::set<std::string> names = abn.GetNames();
-   REQUIRE(abn.Size() == names.size());
+   REQUIRE(abn.Size() == (int)names.size());
 
    for (auto name : names)
    {
